Use "pyramid" shape with different PRNG
If a different PRNG is to be used, it should be easy to spot that the
visual checksum follows a different spec, in order to avoid confusion.

Therefore, v2 or "pyramies" are displayed in a 10 block high pyramid shape,
which is slightly less space efficient and looses 2 bits against the
original scheme (30 vs. 32 individual blocks).